Code refactor: make script more modular
